Update of the request status codes.
.env implementation
