+++ config/constants.js
@@ -1,20 +1,21 @@
+require('dotenv').config();
 module.exports.mongoDbSettings = {
   Settings: {
-    host: 'cluster0-shard-00-00.ampj6.mongodb.net,cluster0-shard-00-01.ampj6.mongodb.net,cluster0-shard-00-02.ampj6.mongodb.net',
-    port: 27017,
-    database: 'test',
-    user: 'jacko',
-    password: 'gj414bDXPeq4fBIt',
+    host: `${process.env.MONGODB_HOST}`,
+    port: `${process.env.MONGODB_PORT}`,
+    database: `${process.env.MONGODB_DB}`,
+    user: `${process.env.MONGODB_USER}`,
+    password: `${process.env.MONGODB_PWD}`,
     ssl: true,
     authSource: 'admin',
   },
 };
+
 module.exports.redisSettings = {
   settings: {
     prefix: '_cache',
-    host: '172.20.0.4',
-    port: 6379,
-    // password: '****'
+    host: `${process.env.REDIS_HOST}`,
+    port: `${process.env.REDIS_PORT}`,
   },
   keyName: 'test:condorlabs-npm-helpers:counter',
 };
